requirements.txt
altair==5.5.0
altgraph @ file:///AppleInternal/Library/BuildRoots/4~CDlvugAeMYaHcTWIELyAQ2joHAecmlI0GAlKPsw/Library/Caches/com.apple.xbs/Sources/python3/altgraph-0.17.2-py2.py3-none-any.whl
attrs==25.4.0
blinker==1.9.0
cachetools==6.2.4
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.1.8
contourpy==1.3.0
cycler==0.12.1
filelock==3.19.1
fonttools==4.60.2
fsspec==2025.10.0
future @ file:///AppleInternal/Library/BuildRoots/4~CDlvugAeMYaHcTWIELyAQ2joHAecmlI0GAlKPsw/Library/Caches/com.apple.xbs/Sources/python3/future-0.18.2-py3-none-any.whl
gitdb==4.0.12
GitPython==3.1.46
idna==3.11
importlib_resources==6.5.2
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.7
macholib @ file:///AppleInternal/Library/BuildRoots/4~CDlvugAeMYaHcTWIELyAQ2joHAecmlI0GAlKPsw/Library/Caches/com.apple.xbs/Sources/python3/macholib-1.15.2-py2.py3-none-any.whl
MarkupSafe==3.0.3
matplotlib==3.9.4
mpmath==1.3.0
narwhals==2.15.0
networkx==3.2.1
numpy==2.0.2
opencv-python==4.12.0.88
packaging==25.0
pandas==2.3.3
pillow==11.3.0
polars==1.36.1
polars-runtime-32==1.36.1
protobuf==6.33.4
psutil==7.2.1
pyarrow==21.0.0
pydeck==0.9.1
pyparsing==3.3.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.1
scipy==1.13.1
six @ file:///AppleInternal/Library/BuildRoots/4~CDlvugAeMYaHcTWIELyAQ2joHAecmlI0GAlKPsw/Library/Caches/com.apple.xbs/Sources/python3/six-1.15.0-py2.py3-none-any.whl
smmap==5.0.2
streamlit==1.50.0
sympy==1.14.0
tenacity==9.1.2
toml==0.10.2
torch==2.8.0
torchvision==0.23.0
tornado==6.5.4
typing_extensions==4.15.0
tzdata==2025.3
ultralytics==8.3.252
ultralytics-thop==2.0.18
urllib3==2.6.3
zipp==3.23.0
